fixing compiler warning
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@23770 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/common/proxybase/RemoteProxyBase.C
@@ -400,7 +400,7 @@ RemoteProxyBase::AddProfileArguments(const MachineProfile &machine,
     if (machine.GetDirectory() != "")
     {
         //Remove all existing instances of -dir, because they're for the LOCAL machine
-        for (int i = 0; i < argv.size(); i++) {
+        for (int i = 0; i < (int) argv.size(); i++) {
             if (argv[i] == "-dir") {
               // First delete the "-dir"
               stringVector::iterator it = argv.erase(argv.begin() + i);
